+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for reasoning models that use "think" as tag for reasoning section
+
 ### Changed
 
 ### Deprecated